refactor: reorganiza os pacotes e cria classes jwttoken
