+++ eazybytes/src/main/java/com/springsecurity/eazybytes/security/filter/token/JWTTokenGeneratorFilter.java
@@ -1,0 +1,4 @@
+package com.springsecurity.eazybytes.security.filter.token;
+
+public class JWTTokenGeneratorFilter {
+}
